backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
motor==3.6.0
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
httpx==0.27.2
openai==1.51.2
google-generativeai==0.8.3
tenacity==9.0.0
